Support for Puppet Group repository (#18)
* Support for Pupper Group repository

Group repository is listed as one of the types. User can add other
Puppet repositories to it. Content is successfully proxies from on of
the sources. Metadata merging was done in simple way, on asumption that
single package will be present only in one repository.

Adding also a integration test to assert proper operation. Test is
designed in that way so Nexus should look at grouped repositories
and merge metadata entries so if module is on second repository it
should be returned.

Co-authored-by: Daniel Sauble <dsauble@sonatype.com>
Co-authored-by: Jeffry Hesse <jhesse@sonatype.com>
Co-authored-by: DarthHater <jeffryxtron@gmail.com>
Co-authored-by: Caio Amaral <camaral@users.noreply.github.com>
Co-authored-by: Jeffry Hesse <5544326+DarthHater@users.noreply.github.com>